DealsBaseFragment implement
DealsBaseFragment have the logic to the
fragmentOffers, fragmentSearch and fragmentFavorite that implement the same logic but now is centralised
